Use babel config and shared processPageDir function.
